move attribute-related methods to an oracle; clean up attribute matcher a bit
