+++ src/oracles/value.rb
@@ -36,4 +36,3 @@ module Oracles
     end
   end
 end
-
